requirements.txt
google-cloud-container-analysis==1.9.0
google-cloud-artifact-registry==1.10.0
google-cloud-logging==3.8.0
atlassian-python-api==3.41.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
Flask==3.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
